Wheel Distance Test
Robot should run forever, outputting its distance and being able to run in a square.

It should run 254 mm (10 inches) in x and y distance.

LMK if you want me to change the distance so that it is manually triggered.  I can also change the while loop so that it only runs 10 laps